+++ src/app/app.routes.ts
@@ -7,20 +7,20 @@ import { ActivityComponent } from '@app/iot/activity/activity.component';
 import { LocationSingleComponent } from '@components/locations/location-single/location-single.component';
 import { GalleryComponent } from '@components/gallery/gallery.component';
 import { DeviceSingleComponent } from '@app/iot/device-single/device-single.component';
-import { DocsComponent } from '@components/docs/docs.component';
+import { DocsComponent } from '@app/iot/docs/docs.component';
 import { Route } from '@angular/router';
 import { LayoutComponent } from './ng5-basic/layout/layout.component';
 import { AuthGuard } from '@services/user.service';
 import { ExperimentalComponent } from '@app/components/experimental/experimental.component';
-import { SendingInformationHttpsComponent } from './components/docs/sending-information-https/sending-information-https.component';
+import { SendingInformationHttpsComponent } from '@app/iot/docs/sending-information-https/sending-information-https.component';
 import { PrivacyPolicyComponent } from '@app/components/privacy-policy/privacy-policy.component';
 import { ProfileComponent } from '@app/components/profile/profile.component';
 import { DeviceEventsActionsComponent } from '@app/iot/device-events-actions/device-events-actions.component';
 import { ContactDetailsComponent } from '@app/components/contact-details/contact-details.component';
 import { GpsComponent } from '@app/iot/gps/gps.component';
-import { RestfulComponent } from '@app/components/docs/restful/restful.component';
+import { RestfulComponent } from '@app/iot/docs/restful/restful.component';
 import { userAuthLayoutRoutes } from '@app/users/user-routes';
-import { AngularCalendarComponent } from '@app/components/angular-calendar/angular-calendar.component';
+import { AngularCalendarComponent } from '@app/boilerplate/angular-calendar/angular-calendar.component';
 
 export function AuthLayoutChild(component: any, route: string, options: any = {}): Route {
   return {
