--- src/app/components/applications-list/applications-list.component.html
@@ -1,3 +1,0 @@
-<div class="application-list-container">
-    <i class="icon-apps"></i>
-</div>
