--- src/app/components/applications-list/applications-list.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-applications-list',
-  templateUrl: './applications-list.component.html',
-  styleUrls: ['./applications-list.component.scss']
-})
-export class ApplicationsListComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
