+++ src/app/iot/docs/docs.component.scss
