+++ src/app/iot/docs/sending-information-https/sending-information-https.component.ts
