+++ src/app/ng5-basic/nav-bar/nav-bar.component.html
@@ -19,9 +19,7 @@
         <div class="d-2">
           <app-notification-list></app-notification-list>
         </div>
-        <!-- <div class="d-2">
-          <app-applications-list></app-applications-list>
-        </div> -->
+
       </div>
     </div>
   </nav>
